implementing the preset list in the frontend
I also added an option skip_required_files to settings.py, and updated 13 of the games to use this setting to skip the error message about the ROM file being missing.

I'm including presets for just four of the games for now.  I'll wait until the formatting is finalized and other fixes are made before uploading the full set of presets.